+++ pyneqsys/tests/test_pysym.py
@@ -15,7 +15,7 @@ def test_pysym_SymbolicSys_from_callback():
         symarray=pysym.symarray,
         Matrix=pysym.Matrix)
 
-    x, sol = ss.solve('scipy', [1, 0], [3])
+    x, sol = ss.solve([1, 0], [3], solver='scipy')
     assert sol['success']
     assert abs(x[0] - 0.8411639) < 2e-7
     assert abs(x[1] - 0.1588361) < 2e-7